[AutoDiff] [AST] Handle null decl in 'printDifferentiableAttrArguments'.
`DeclAttribute::print()` accepts null decls and has a null default argument.  A segfault was triggered in SourceKitService because `printDifferentiableAttrArguments()` took a null decl and deferenced it.  This patch makes `printDifferentiableAttrArguments()` handle null decls — if the decl is null, we do not print the differentiability parameters or the `where` clause.

I have not added a unit test however because I haven't been able to reproduce it independently.  This will be done later as this change is low-risk.

Resolves rdar://76095289.