Record initial parent map
We need to be able to traverse back to the top of expressions in order
to emit the diagnostic to the right place. This patch records the parent
map of the highest expression in the tree that the checker was
constructed with. This will record the mapping for all sub-expressions
as well.