[Tests] Do not ignore swiftsourceinfo by default in lit tests
This was added as a default in b103997c293a55e9d42cd659b840dfc1fdf51070,
with the reasoning that users do not have .swiftsourceinfo for the
stdlib. A later change (51d6243d2451dfb0a05996f902e6766834f1cad7) then
avoids emitting .swiftsourceinfo for the stdlib at all, so this is no
longer an issue.