[CMake] Fix some stdlib dependencies dropping architectures
There were a couple situations where we had dependencies like `target-arch -> dependency` in the stdlib that would drop the architecture, thus causing unnecessary architecture to get built.

For example `swift-reflection-test-watchsimulator-x86_64` used to depend on `swiftSwiftOnoneSupport-watchsimulator` but should depend on `swiftSwiftOnoneSupport-watchsimulator-x86_64`.

Avoid dropping the architecture in the dependencies to fix the problem.