Adding initial diagnostic collector impl
I'm adding an initial implementation for the diagnostic collection
mechanism. This is sufficient to remove the ad-hoc collection and
diagnosing that I was doing in commit
d0d9bef703e549aa112393bf76882c3b1ca146f7.

This won't be quite be sufficient for the needs of the constraint
solver. We will need to keep some additional context around to keep the
Solution object available.